moved files, updated file path, moved Landing into own function.
