+++ src/landing/quote/Quote.jsx
